Fix deadlock, add settings_local support
