+++ .gitignore
@@ -4,3 +4,4 @@ logs/
 *.env
 NOTES
 tmp/
+settings_local.py
